fix(shared): address aiAct verbose review feedback
