+++ packages/shared/src/cli/verbose.ts
@@ -1,5 +1,10 @@
-import { basename, dirname, isAbsolute, join, relative } from 'node:path';
-import { writeCliScreenshotFile } from './screenshot-file';
+import { type CliVerboseLine, buildAiActTimelineLines } from './verbose-ai-act';
+import {
+  type CliVerboseScreenshotCollectOptions,
+  type CliVerboseScreenshotExportMode,
+  collectScreenshotRefs,
+  renderScreenshotList,
+} from './verbose-screenshot';
 
 export const cliVerboseFlag = 'verbose';
 
@@ -33,14 +38,6 @@ type DumpUpdateAgent = {
   ) => () => void;
   reportFile?: string | null;
 };
-
-interface CliVerboseScreenshotRefLike {
-  type: 'midscene_screenshot_ref';
-  id?: unknown;
-  mimeType?: unknown;
-  storage?: unknown;
-  path?: string;
-}
 
 interface CliVerboseExecutionTaskLike {
   taskId?: unknown;
@@ -80,11 +77,6 @@ interface CliVerboseExecutionDumpLike {
   name?: unknown;
   description?: unknown;
   tasks: CliVerboseExecutionTaskLike[];
-}
-
-interface CliVerboseLine {
-  key: string;
-  text: string;
 }
 
 const getGlobalContext = (): { current: CliVerboseContext } => {
@@ -434,148 +426,6 @@ function summarizeTaskText(task: CliVerboseExecutionTaskLike): string {
   return detail ? `${label} ${status}: ${detail}` : `${label} ${status}`;
 }
 
-function isCliVerboseScreenshotRefLike(
-  value: unknown,
-): value is CliVerboseScreenshotRefLike {
-  return isRecord(value) && value.type === 'midscene_screenshot_ref';
-}
-
-function toSerializableScreenshot(
-  value: unknown,
-): CliVerboseScreenshotRefLike | null {
-  if (!value || typeof value !== 'object') {
-    return null;
-  }
-
-  const maybeSerializable = value as {
-    toSerializable?: () => unknown;
-  };
-  if (typeof maybeSerializable.toSerializable === 'function') {
-    try {
-      const serialized = maybeSerializable.toSerializable();
-      return isCliVerboseScreenshotRefLike(serialized) ? serialized : null;
-    } catch {
-      return null;
-    }
-  }
-
-  return isCliVerboseScreenshotRefLike(value) ? value : null;
-}
-
-function getStringProperty(value: unknown, key: string): string | undefined {
-  if (!isRecord(value)) {
-    return undefined;
-  }
-
-  try {
-    const property = value[key];
-    return typeof property === 'string' && property.length > 0
-      ? property
-      : undefined;
-  } catch {
-    return undefined;
-  }
-}
-
-function screenshotRawBase64(value: unknown): string | undefined {
-  const rawBase64 = getStringProperty(value, 'rawBase64');
-  if (rawBase64) {
-    return rawBase64;
-  }
-
-  const base64 = getStringProperty(value, 'base64');
-  const match = base64?.match(/^data:image\/(?:png|jpeg|jpg);base64,(.+)$/);
-  return match?.[1];
-}
-
-function exportInlineScreenshotForVerbose(
-  value: unknown,
-  serialized: CliVerboseScreenshotRefLike,
-  reportFile?: unknown,
-): string | undefined {
-  if (typeof serialized.path === 'string') {
-    return serialized.path;
-  }
-
-  const rawBase64 = screenshotRawBase64(value);
-  if (!rawBase64) {
-    return undefined;
-  }
-
-  try {
-    const directoryPath =
-      typeof reportFile === 'string' && reportFile.length > 0
-        ? join(dirname(reportFile), 'screenshots')
-        : undefined;
-    return writeCliScreenshotFile(rawBase64, {
-      id: serialized.id,
-      mimeType: serialized.mimeType,
-      extension: getStringProperty(value, 'extension'),
-      ...(directoryPath
-        ? { directoryPath }
-        : { directoryName: 'midscene-cli-screenshots' }),
-      overwrite: false,
-    });
-  } catch {
-    return undefined;
-  }
-}
-
-function collectScreenshotRefs(
-  value: unknown,
-  reportFile?: unknown,
-): Array<Record<string, unknown>> {
-  const screenshots: Array<Record<string, unknown>> = [];
-  const visit = (candidate: unknown, timing?: unknown): void => {
-    if (Array.isArray(candidate)) {
-      for (const item of candidate) {
-        visit(item, timing);
-      }
-      return;
-    }
-
-    const serialized = toSerializableScreenshot(candidate);
-    if (serialized?.type === 'midscene_screenshot_ref') {
-      const screenshot: Record<string, unknown> = {
-        id: serialized.id,
-        storage: serialized.storage,
-      };
-      const exportedPath = exportInlineScreenshotForVerbose(
-        candidate,
-        serialized,
-        reportFile,
-      );
-      if (exportedPath) {
-        screenshot.path = exportedPath;
-        screenshot.file = basename(exportedPath);
-      }
-      if (typeof timing === 'string') {
-        screenshot.timing = timing;
-      }
-      screenshots.push(screenshot);
-      return;
-    }
-
-    if (!isRecord(candidate)) {
-      return;
-    }
-
-    const screenshotRecord = candidate.screenshot;
-    if (screenshotRecord) {
-      visit(screenshotRecord, candidate.timing);
-    }
-    if (candidate.recorder) {
-      visit(candidate.recorder);
-    }
-    if (isRecord(candidate.uiContext)) {
-      visit(candidate.uiContext.screenshot);
-    }
-  };
-
-  visit(value);
-  return screenshots;
-}
-
 function isCliVerboseExecutionDumpLike(
   value: unknown,
 ): value is CliVerboseExecutionDumpLike {
@@ -584,6 +434,7 @@ function isCliVerboseExecutionDumpLike(
 
 function summarizeDumpUpdate(
   executionDump: unknown,
+  screenshotOptions: CliVerboseScreenshotCollectOptions = {},
 ): Record<string, unknown> | undefined {
   if (!isCliVerboseExecutionDumpLike(executionDump)) {
     return undefined;
@@ -603,7 +454,7 @@ function summarizeDumpUpdate(
       message: summarizeTaskText(task),
       error: task.errorMessage,
       durationMs: task.timing?.cost,
-      screenshots: collectScreenshotRefs(task).slice(-3),
+      screenshots: collectScreenshotRefs(task, screenshotOptions).slice(-3),
     }),
   );
 
@@ -629,7 +480,7 @@ function summarizeDumpUpdate(
           durationMs: latestTask.timing?.cost,
         }
       : undefined,
-    screenshots: collectScreenshotRefs(tasks).slice(-5),
+    screenshots: collectScreenshotRefs(tasks, screenshotOptions).slice(-5),
   };
 }
 
@@ -640,31 +491,6 @@ function stringifyArgs(args: unknown): string {
 
   return Object.entries(args)
     .map(([key, value]) => `${key}=${compactText(value)}`)
-    .join(', ');
-}
-
-function renderScreenshotList(screenshots: unknown): string {
-  if (!Array.isArray(screenshots) || screenshots.length === 0) {
-    return '';
-  }
-
-  return screenshots
-    .map((item) => {
-      if (!isRecord(item)) {
-        return '';
-      }
-      const path =
-        typeof item.path === 'string'
-          ? item.path
-          : typeof item.file === 'string'
-            ? item.file
-            : typeof item.id === 'string'
-              ? item.id
-              : '';
-      const timing = typeof item.timing === 'string' ? item.timing : '';
-      return [timing, path].filter(Boolean).join(' ');
-    })
-    .filter(Boolean)
     .join(', ');
 }
 
@@ -695,482 +521,6 @@ function renderLinesOnce(
   return pendingLines.length > 0
     ? pendingLines.map((line) => line.text).join('\n')
     : undefined;
-}
-
-function taskKey(task: CliVerboseExecutionTaskLike, fallback: string): string {
-  return typeof task.taskId === 'string' && task.taskId.length > 0
-    ? task.taskId
-    : fallback;
-}
-
-function numberFromUnknown(value: unknown): number | undefined {
-  return typeof value === 'number' && Number.isFinite(value)
-    ? value
-    : undefined;
-}
-
-function integerText(value: number): string {
-  return String(Math.round(value));
-}
-
-function formatPoint(point: readonly [number, number]): string {
-  return `(${integerText(point[0])}, ${integerText(point[1])})`;
-}
-
-function formatBbox(bbox: readonly [number, number, number, number]): string {
-  return `(${bbox.map(integerText).join(',')})`;
-}
-
-function bboxArrayFromProperty(
-  value: Record<string, unknown>,
-  key: string,
-): [number, number, number, number] | undefined {
-  const bbox = value[key];
-  if (!Array.isArray(bbox) || bbox.length < 4) {
-    return undefined;
-  }
-
-  const left = numberFromUnknown(bbox[0]);
-  const top = numberFromUnknown(bbox[1]);
-  const right = numberFromUnknown(bbox[2]);
-  const bottom = numberFromUnknown(bbox[3]);
-  if (
-    left === undefined ||
-    top === undefined ||
-    right === undefined ||
-    bottom === undefined
-  ) {
-    return undefined;
-  }
-
-  return [left, top, right, bottom];
-}
-
-function centerPointFromBbox(
-  bbox: readonly [number, number, number, number],
-): [number, number] {
-  return [
-    Math.floor((bbox[0] + bbox[2]) / 2),
-    Math.floor((bbox[1] + bbox[3]) / 2),
-  ];
-}
-
-function pathForReportScreenshot(path: string, reportFile?: unknown): string {
-  const resolvedPath =
-    typeof reportFile === 'string' &&
-    reportFile.length > 0 &&
-    (path.startsWith('./') || path.startsWith('../'))
-      ? join(dirname(reportFile), path)
-      : path;
-
-  if (!isAbsolute(resolvedPath)) {
-    return resolvedPath;
-  }
-
-  const relativePath = relative(process.cwd(), resolvedPath);
-  if (
-    relativePath &&
-    !relativePath.startsWith('..') &&
-    !isAbsolute(relativePath)
-  ) {
-    return relativePath;
-  }
-
-  return resolvedPath;
-}
-
-function latestScreenshotPathForAiAct(
-  value: unknown,
-  reportFile?: unknown,
-): string {
-  const screenshot = collectScreenshotRefs(value, reportFile)
-    .slice()
-    .reverse()
-    .find((item) => typeof item.path === 'string' && item.path.length > 0);
-  const path = typeof screenshot?.path === 'string' ? screenshot.path : '';
-  return path ? pathForReportScreenshot(path, reportFile) : '';
-}
-
-function planLimitText(task: CliVerboseExecutionTaskLike): string {
-  if (!isRecord(task.param)) {
-    return '';
-  }
-  const limit = numberFromUnknown(task.param.replanningCycleLimit);
-  return limit && limit > 0 ? `/${integerText(limit)}` : '';
-}
-
-function planPrefix(
-  task: CliVerboseExecutionTaskLike,
-  planIndex: number,
-): string {
-  return `[Midscene][aiAct][Plan ${planIndex}${planLimitText(task)}]`;
-}
-
-function actionOutputText(value: unknown): string {
-  if (!Array.isArray(value) || value.length === 0) {
-    return '';
-  }
-
-  const action = value.find(isRecord);
-  if (!action) {
-    return '';
-  }
-
-  return (
-    compactText(action.log) ||
-    compactText(action.thought) ||
-    compactText(summarizeParam(action.param))
-  );
-}
-
-function plannedTextForAiAct(task: CliVerboseExecutionTaskLike): string {
-  if (!isRecord(task.output)) {
-    return compactText(task.output) || compactText(task.thought);
-  }
-
-  return (
-    compactText(task.output.log) ||
-    compactText(task.output.thought) ||
-    compactText(task.thought) ||
-    summarizeSubGoals(task.output.updateSubGoals) ||
-    actionOutputText(task.output.actions) ||
-    compactText(task.output.output)
-  );
-}
-
-function completeTextForAiAct(task: CliVerboseExecutionTaskLike): string {
-  if (!isRecord(task.output)) {
-    return '';
-  }
-
-  const output = compactText(task.output.output);
-  if (output) {
-    return output;
-  }
-
-  return task.output.shouldContinuePlanning === false
-    ? plannedTextForAiAct(task)
-    : '';
-}
-
-function pointFromLocateLike(
-  value: Record<string, unknown>,
-): [number, number] | undefined {
-  const center = value.center;
-  if (Array.isArray(center) && center.length >= 2) {
-    const x = numberFromUnknown(center[0]);
-    const y = numberFromUnknown(center[1]);
-    if (x !== undefined && y !== undefined) {
-      return [x, y];
-    }
-  }
-
-  const point = value.point;
-  if (Array.isArray(point) && point.length >= 2) {
-    const x = numberFromUnknown(point[0]);
-    const y = numberFromUnknown(point[1]);
-    if (x !== undefined && y !== undefined) {
-      return [x, y];
-    }
-  }
-
-  const locatedPixelBbox = bboxArrayFromProperty(value, 'locatedPixelBbox');
-  if (locatedPixelBbox) {
-    return centerPointFromBbox(locatedPixelBbox);
-  }
-
-  const bbox = bboxArrayFromProperty(value, 'bbox');
-  if (bbox) {
-    return centerPointFromBbox(bbox);
-  }
-
-  return undefined;
-}
-
-function bboxFromLocateLike(
-  value: Record<string, unknown>,
-): [number, number, number, number] | undefined {
-  const locatedPixelBbox = bboxArrayFromProperty(value, 'locatedPixelBbox');
-  if (locatedPixelBbox) {
-    return locatedPixelBbox;
-  }
-
-  const bbox = bboxArrayFromProperty(value, 'bbox');
-  if (bbox) {
-    return bbox;
-  }
-
-  if (!isRecord(value.rect)) {
-    return undefined;
-  }
-
-  const left = numberFromUnknown(value.rect.left);
-  const top = numberFromUnknown(value.rect.top);
-  const width = numberFromUnknown(value.rect.width);
-  const height = numberFromUnknown(value.rect.height);
-  if (
-    left === undefined ||
-    top === undefined ||
-    width === undefined ||
-    height === undefined
-  ) {
-    return undefined;
-  }
-
-  return [left, top, left + width, top + height];
-}
-
-function targetTextFromLocateLike(value: Record<string, unknown>): string {
-  return (
-    compactText(value.description) ||
-    compactText(value.prompt) ||
-    summarizeUserInstruction(value)
-  );
-}
-
-function isLocateLike(value: unknown): value is Record<string, unknown> {
-  if (!isRecord(value)) {
-    return false;
-  }
-
-  return (
-    'center' in value ||
-    'rect' in value ||
-    'point' in value ||
-    'bbox' in value ||
-    'prompt' in value ||
-    'description' in value
-  );
-}
-
-const locatorParamKeys = ['locate', 'from', 'to', 'start', 'end'];
-
-function firstLocateLikeParam(
-  param: unknown,
-): Record<string, unknown> | undefined {
-  if (!isRecord(param)) {
-    return undefined;
-  }
-
-  for (const key of locatorParamKeys) {
-    const value = param[key];
-    if (isLocateLike(value)) {
-      return value;
-    }
-  }
-
-  return Object.values(param).find(isLocateLike);
-}
-
-function hasUnresolvedLocateLikeParam(param: unknown): boolean {
-  if (!isRecord(param)) {
-    return false;
-  }
-
-  return Object.entries(param).some(([key, value]) => {
-    if (locatorParamKeys.includes(key) && typeof value === 'string') {
-      return true;
-    }
-
-    return (
-      isLocateLike(value) &&
-      !pointFromLocateLike(value) &&
-      !bboxFromLocateLike(value)
-    );
-  });
-}
-
-interface AiActActionText {
-  action: string;
-  running: string;
-  done: string;
-}
-
-function sleepActionText(
-  actionName: string,
-  param: unknown,
-): AiActActionText | undefined {
-  if (actionName !== 'Sleep' || !isRecord(param)) {
-    return undefined;
-  }
-
-  const timeMs =
-    numberFromUnknown(param.timeMs) ??
-    numberFromUnknown(param.duration) ??
-    numberFromUnknown(param.timeoutMs);
-  const action = timeMs ? `Sleep ${integerText(timeMs)}ms` : 'Sleep';
-  return {
-    action,
-    running: action,
-    done: 'Sleep',
-  };
-}
-
-function locateActionText(
-  actionName: string,
-  param: unknown,
-): AiActActionText | undefined {
-  const locate = firstLocateLikeParam(param);
-  if (!locate) {
-    return undefined;
-  }
-
-  const point = pointFromLocateLike(locate);
-  const bbox = bboxFromLocateLike(locate);
-  if (!point) {
-    return undefined;
-  }
-
-  const target = targetTextFromLocateLike(locate);
-  const targetSegment = target ? ` "${target}"` : '';
-  const pointSegment = ` at ${formatPoint(point)}`;
-  const bboxSegment = bbox ? `, bbox=${formatBbox(bbox)}` : '';
-  return {
-    action: `${actionName}${targetSegment}${pointSegment}${bboxSegment}`,
-    running: `${actionName} at ${formatPoint(point)}`,
-    done: actionName,
-  };
-}
-
-function genericActionText(
-  actionName: string,
-  param: unknown,
-): AiActActionText | undefined {
-  if (hasUnresolvedLocateLikeParam(param)) {
-    return undefined;
-  }
-
-  const paramText = compactText(summarizeParam(param));
-  const action = paramText ? `${actionName}: ${paramText}` : actionName;
-  return {
-    action,
-    running: action,
-    done: actionName,
-  };
-}
-
-function actionTextForAiAct(
-  task: CliVerboseExecutionTaskLike,
-): AiActActionText | undefined {
-  const actionName =
-    typeof task.subType === 'string' && task.subType.length > 0
-      ? task.subType
-      : 'Action';
-  return (
-    sleepActionText(actionName, task.param) ||
-    locateActionText(actionName, task.param) ||
-    genericActionText(actionName, task.param)
-  );
-}
-
-function buildAiActTimelineLines(
-  executionDump: unknown,
-  reportFile?: unknown,
-): CliVerboseLine[] {
-  if (!isCliVerboseExecutionDumpLike(executionDump)) {
-    return [];
-  }
-
-  const lines: CliVerboseLine[] = [];
-  let planIndex = 0;
-  let currentPlan:
-    | {
-        index: number;
-        prefix: string;
-      }
-    | undefined;
-
-  executionDump.tasks.forEach((task, taskIndex) => {
-    const fallbackKey = String(taskIndex + 1);
-    if (task.type === 'Planning' && task.subType === 'Plan') {
-      planIndex += 1;
-      currentPlan = {
-        index: planIndex,
-        prefix: planPrefix(task, planIndex),
-      };
-      const keyPrefix = `aiAct:plan:${taskKey(task, fallbackKey)}`;
-      const screenshotPath = latestScreenshotPathForAiAct(task, reportFile);
-      if (screenshotPath) {
-        lines.push({
-          key: `${keyPrefix}:thinking`,
-          text: `${currentPlan.prefix} Thinking with the latest screenshot: ${screenshotPath}`,
-        });
-      }
-
-      if (task.status === 'finished') {
-        const plannedText = plannedTextForAiAct(task);
-        if (plannedText) {
-          lines.push({
-            key: `${keyPrefix}:planned`,
-            text: `${currentPlan.prefix} Planned: ${plannedText}`,
-          });
-        }
-
-        const completeText = completeTextForAiAct(task);
-        if (completeText) {
-          lines.push({
-            key: `${keyPrefix}:complete`,
-            text: `[Midscene][aiAct] Complete: ${completeText}`,
-          });
-        }
-      }
-      return;
-    }
-
-    if (
-      !currentPlan ||
-      task.type !== 'Action Space' ||
-      task.subType === 'Finished'
-    ) {
-      return;
-    }
-
-    const actionText = actionTextForAiAct(task);
-    if (!actionText) {
-      return;
-    }
-
-    const keyPrefix = `aiAct:action:${taskKey(task, fallbackKey)}`;
-    lines.push({
-      key: `${keyPrefix}:planned`,
-      text: `${currentPlan.prefix} Action: ${actionText.action}`,
-    });
-
-    if (task.status === 'running') {
-      lines.push({
-        key: `${keyPrefix}:running`,
-        text: `[Midscene][aiAct][Action] Running: ${actionText.running}`,
-      });
-    }
-
-    if (task.status === 'finished') {
-      const cost =
-        typeof task.timing?.cost === 'number'
-          ? ` cost=${integerText(task.timing.cost)}ms`
-          : '';
-      lines.push({
-        key: `${keyPrefix}:finished`,
-        text: `[Midscene][aiAct][Action] Done: ${actionText.done}${cost}`,
-      });
-    }
-
-    if (task.status === 'failed') {
-      const cost =
-        typeof task.timing?.cost === 'number'
-          ? ` cost=${integerText(task.timing.cost)}ms`
-          : '';
-      const error =
-        typeof task.errorMessage === 'string' && task.errorMessage.length > 0
-          ? ` error=${task.errorMessage}`
-          : '';
-      lines.push({
-        key: `${keyPrefix}:failed`,
-        text: `[Midscene][aiAct][Action] Failed: ${actionText.done}${cost}${error}`,
-      });
-    }
-  });
-
-  return lines;
 }
 
 function renderCliVerboseEventText(
@@ -1279,6 +629,18 @@ function renderCliVerboseEventText(
     }
     case 'command_done': {
       if (isActVerboseEvent(command, tool)) {
+        if (event.status === 'error') {
+          const error =
+            typeof event.error === 'string' && event.error.length > 0
+              ? `: ${event.error}`
+              : '';
+          return renderLinesOnce(context, [
+            {
+              key: `aiAct:command_failed:${event.error ?? 'unknown'}`,
+              text: `[Midscene][aiAct] Failed${error}`,
+            },
+          ]);
+        }
         return undefined;
       }
       const status = event.status === 'error' ? 'failed' : 'finished';
@@ -1306,15 +668,27 @@ export function attachCliVerboseDumpListener(
     return () => {};
   }
 
+  const screenshotExportCache = new Map<string, string>();
   return agent.addDumpUpdateListener((_dump, executionDump) => {
-    const summary = summarizeDumpUpdate(executionDump);
+    const context = getCliVerboseContext();
+    const isTextMode = context.format !== 'jsonl';
+    const isActTool = options?.toolName === 'act';
+    const summaryExportMode: CliVerboseScreenshotExportMode =
+      isTextMode && !isActTool ? 'tmp' : 'none';
+    const summary = summarizeDumpUpdate(executionDump, {
+      exportMode: summaryExportMode,
+      cache: screenshotExportCache,
+    });
     if (!summary) {
       return;
     }
-    const context = getCliVerboseContext();
     const aiActTimeline =
-      context.format !== 'jsonl' && options?.toolName === 'act'
-        ? buildAiActTimelineLines(executionDump, agent.reportFile)
+      isTextMode && isActTool
+        ? buildAiActTimelineLines(executionDump, {
+            reportFile: agent.reportFile,
+            exportMode: 'report',
+            cache: screenshotExportCache,
+          })
         : undefined;
     emitCliVerboseEvent({
       event: 'dump_update',
